:sparkles: Add codon scoring
Because this wasn't stupid enough already. And I was a bit bored. Holidays
can do that to you. That and I didn't want to spend all day watching TV, and
I was making some bread and needed to do something while the dough sat for a
couple of hours.

Am I over-sharing?

Sorry.